+++ steelc/src/building/project_builder.cpp
@@ -23,9 +23,11 @@
 #include "../stproj/bad_stproj_exception.h"
 #include "../output/output.h"
 #include "../error/error_printer.h"
+#include "../codegen/ir/ir_holder.h"
 
 bool project_builder::load_project(const std::string& project_path) {
 	// load stproj
+	output::verbose("Loading project file at: \'{}\'\n", console_colors::DIM, project_path);
 	try {
 		project_file = stproj_file::load(project_path);
 	}
@@ -39,42 +41,38 @@ bool project_builder::build_project() {
 	mark_build_start();
 	output::print("Build started...\n");
 
-	// load build cache
-	build_cache_file cache = load_cache();
-	std::vector<source_file> to_compile = get_files_to_compile(cache);
+	// compile project
+	std::vector<ir_holder> generated_ir = compile_project();
 
-	// compile sources
-	compiler compiler(to_compile);
-	compile_config compile_cfg{}; // TODO: generate config based on project settings + cl args
-
-	mark_compile_start();
-	if (compiler.compile(compile_cfg)) {
-		output::print("Compilation succeeded. ", console_colors::BOLD + console_colors::GREEN);
-		output::print("(Took {:.3f} seconds)\n", console_colors::DIM, get_compilation_time());
-
-		// save build cache
-		save_cache(cache);
-	}
-	else {
-		output::err("Compilation failed with {} errors.\n", console_colors::BOLD + console_colors::RED, compiler.get_error_count());
-		error_printer::print_errors(compiler.get_errors());
-		return false;
-	}
-
-	// output code
-	outputter = std::make_unique<code_outputter>(project_file->parent_path().string(), build_cfg);
-
-	for (const auto& il_h : compiler.get_generated_ir()) {
-		std::string path = get_ir_path(*il_h.owning_unit->source_file);
-		if (outputter->output_il(il_h.ir, path) != OUTPUT_SUCCESS) {
-			output::err("Failed to output IR file: {}\n", console_colors::BOLD + console_colors::RED, path);
+	// output code (if any)
+	if (generated_ir.size() > 0) {
+		output::verbose("Initializing output system...\n");
+		output::verbose("Output directory: {}\n", console_colors::DIM, build_cfg.output_dir);
+		output::verbose("Intermediate directory: {}\n", console_colors::DIM, build_cfg.intermediate_dir);
+		outputter = code_outputter::create(project_file->parent_path().string(), build_cfg);
+		if (!outputter) {
+			output::err("Failed to initialize output system, ensure output paths are valid and accessable.\n", console_colors::BOLD + console_colors::RED);
 			return false;
+		}
+
+		for (const auto& il_h : generated_ir) {
+			std::string path = get_ir_path(*il_h.owning_unit->source_file);
+			if (outputter->output_il(il_h.ir, path) != OUTPUT_SUCCESS) {
+				output::err("Failed to output IR file: {}\n", console_colors::BOLD + console_colors::RED, path);
+				return false;
+			}
 		}
 	}
 
 	// gather all irs
 	for (auto& src : project_file->sources) {
 		all_irs.push_back(get_ir_path(src, false));
+	}
+
+	// build with clang
+	if (!clang_build(all_irs)) {
+		output::err("Clang: building failed.\n", console_colors::BOLD + console_colors::RED);
+		return false;
 	}
 
 	output::print("Building succeeded. ", console_colors::BOLD + console_colors::GREEN);
@@ -91,10 +89,48 @@ bool project_builder::build_project() {
 
 	return true;
 }
+std::vector<ir_holder> project_builder::compile_project() {
+	// load build cache
+	build_cache_file cache = load_cache();
+	std::vector<source_file> to_compile;
+
+	if (build_cfg.build_all) {
+		to_compile = project_file->sources;
+	}
+	else {
+		to_compile = get_files_to_compile(cache);
+	}
+
+	// early out if nothing to compile
+	if (to_compile.empty()) {
+		output::print("No changes detected, skipping compilation.\n", console_colors::BLUE);
+		return {};
+	}
+
+	// compile sources
+	compiler cmp = compiler(to_compile);
+	compile_config compile_cfg{}; // TODO: generate config based on project settings + cl args
+
+	mark_compile_start();
+	output::print("Compiling {} source file(s)...\n", console_colors::BOLD, to_compile.size());
+	if (cmp.compile(compile_cfg)) {
+		output::print("Compilation succeeded. ", console_colors::BOLD + console_colors::GREEN);
+		output::print("(Took {:.3f} seconds)\n", console_colors::DIM, get_compilation_time());
+
+		// save build cache
+		save_cache(cache);
+	}
+	else {
+		output::err("Compilation failed with {} errors.\n", console_colors::BOLD + console_colors::RED, cmp.get_error_count());
+		error_printer::print_errors(cmp.get_errors());
+		return {};
+	}
+
+	return cmp.get_generated_ir();
+}
 
 int project_builder::run_build_command(const std::string& command) {
 	std::string replaced = replace_vars(command);
-	output::verbose("Replaced build command vars: {}\n", console_colors::DIM, replaced);
 	return system(replaced.c_str());
 }
 
@@ -108,18 +144,22 @@ double project_builder::get_compilation_time() const {
 std::string project_builder::get_ir_path(const source_file& src, bool relative) const {
 	std::string rel_path = "./IL/" + src.relative_path + ".ll";
 	if (!relative) {
-		return (outputter->get_intermediate_dir() / rel_path).string();
+		return (get_intermediate_dir() / rel_path).string();
 	}
 	return rel_path;
 }
 
 build_cache_file project_builder::load_cache() {
 	cache_path = project_file->parent_path() / build_cfg.build_cache_file;
+	output::verbose("Loading build cache at: \'{}\'\n", console_colors::DIM, cache_path.string());
 	// load existing cache
 	if (std::filesystem::exists(cache_path)) {
 		build_cache_file cache;
 		if (cache.deserialize(cache_path)) {
 			return cache;
+		}
+		else {
+			output::print("Warn: failed to load build cache, using default.\n", console_colors::YELLOW);
 		}
 	}
 
@@ -176,10 +216,14 @@ std::vector<source_file> project_builder::get_files_to_compile(build_cache_file&
 		source_set.insert(src.full_path);
 	}
 
+	std::vector<std::string> to_remove;
 	for (const auto& [path, meta] : metadata) {
 		if (!source_set.contains(path)) {
-			cache.remove_metadata(path);
+			to_remove.push_back(path);
 		}
+	}
+	for (const auto& path : to_remove) {
+		cache.remove_metadata(path);
 	}
 
 	return out_files;
@@ -194,6 +238,25 @@ file_metadata project_builder::generate_metadata(const source_file& src) {
 	meta.hash = file_hasher::hash_file(src.full_path, &meta.size);
 
 	return meta;
+}
+
+bool project_builder::clang_build(const std::vector<std::string>& ir_files) {
+	std::string cmd = "clang ";
+	for (const auto& ir : ir_files) {
+		cmd += "\"" + ir + "\" ";
+	}
+	cmd += "-o \"" + (get_output_dir() / (project_filename() + get_platform_app_extension())).string() + "\"";
+	return system(cmd.c_str()) == 0;
+}
+
+inline std::string project_builder::get_platform_app_extension() const {
+#if defined(_WIN32) || defined(_WIN64)
+	return ".exe";
+#elif defined(__APPLE__) || defined(__linux__)
+	return ".out";
+#else
+	return "";
+#endif
 }
 
 std::string project_builder::replace_vars(const std::string& str) {
@@ -213,7 +276,7 @@ std::string project_builder::replace_vars(const std::string& str) {
 			}
 		}
 		else if (key == "OUTPUT_DIR") {
-			out += outputter->get_output_dir().string();
+			out += get_output_dir().string();
 		}
 		
 		remaining = m.suffix().str();