+++ steelc/src/building/project_builder.h
@@ -10,8 +10,10 @@
 #include "cache/build_cache_file.h"
 #include "build_config.h"
 #include "code_outputter.h"
+#include "../compiler.h"
 #include "../stproj/source_file.h"
 #include "../stproj/stproj_file.h"
+#include "../codegen/ir/ir_holder.h"
 
 class project_builder {
 public:
@@ -21,6 +23,7 @@ public:
 
 	bool load_project(const std::string& project_path);
 	bool build_project();
+	std::vector<ir_holder> compile_project();
 	
 	int run_build_command(const std::string& command);
 
@@ -55,11 +58,22 @@ private:
 
 	std::string get_ir_path(const source_file& src, bool relative = true) const;
 
+	inline std::filesystem::path get_output_dir() const {
+		return project_file->parent_path() / build_cfg.output_dir;
+	}
+	inline std::filesystem::path get_intermediate_dir() const {
+		return project_file->parent_path() / build_cfg.intermediate_dir;
+	}
+
 	// build cache
 	build_cache_file load_cache();
 	void save_cache(const build_cache_file& cache) const;
 	std::vector<source_file> get_files_to_compile(build_cache_file& cache);
 	file_metadata generate_metadata(const source_file& src);
+
+	bool clang_build(const std::vector<std::string>& ir_files);
+
+	inline std::string get_platform_app_extension() const;
 
 	std::string replace_vars(const std::string& str);
 };