+++ steelc/src/steelc_definitions.h
@@ -1,3 +1,3 @@
 #pragma once
 
-#define STEELC_VERSION_STR "0.3.0-alpha";
+#define STEELC_VERSION_STR "0.4.0-alpha";