+++ steelc/src/utils/path_utils.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include <string>
+#include <filesystem>
+
+namespace path_utils {
+	std::filesystem::path normalize_path(const std::string& path);
+}